[SPARK-36895][SQL] Add Create Index syntax support
### What changes were proposed in this pull request?
This is the 2nd PR for DSv2 index support.

This PR adds the following:

- create index syntax support in parser and analyzer
- `CreateIndex` logic node
- `CreateIndexExec` physical node

`CreateIndex` is not implemented yet in this PR. Calling `CreateIndex` will throw `SQLFeatureNotSupportedException`, and the parsed index information such as `IndexName` `indexType` `columns` and index properties will be included in the error message for now for testing purpose.

### Why are the changes needed?
To support index in DSv2

### Does this PR introduce _any_ user-facing change?
Yes, the create table syntax as the following:

```
CREATE INDEX index_name ON [TABLE] table_name [USING index_type] (column_index_property_list)[OPTIONS indexPropertyList]

    column_index_property_list: column_name [OPTIONS(indexPropertyList)]  [ ,  . . . ]
    indexPropertyList: index_property_name [= index_property_value] [ ,  . . . ]
```

### How was this patch tested?
add a UT

Closes #34148 from huaxingao/index_syntax.

Lead-authored-by: Huaxin Gao <huaxin_gao@apple.com>
Co-authored-by: Huaxin Gao <huaxin.gao11@gmail.com>
Signed-off-by: DB Tsai <d_tsai@apple.com>