[SPARK-39074][CI] Fail on upload, not download of missing test files
### What changes were proposed in this pull request?
The CI should not fail when there are no test result files to be downloaded.

### Why are the changes needed?
The CI workflow "Report test results" fails when there are no artifacts to be downloaded from the triggering workflow. In some situations, the triggering workflow is not skipped, but all test jobs are skipped in case no code changes are detected.

Examples:
- https://github.com/G-Research/spark/runs/6208072225?check_suite_focus=true#step:4:14
- https://github.com/G-Research/spark/runs/6094719451?check_suite_focus=true#step:4:14

In that situation, no test files are uploaded, which makes the triggered workflow fail.

Downloading no test files can have two reasons:
1. The code has been build but no tests have been executed, or tests have been executed but no test files have been generated.
3. No code has been built and tested deliberately.

You want to be notified in the first situation to fix the CI. Therefore, CI should fail when code is built and tests are run but no test result files are been found.

The `dawidd6/action-download-artifact` action cannot be configured to not fail when there are no files. But Github Actions can be configured (`continue-on-error`) to ignore that and still have the job succeed. However, if that action fails for a different reason, this will not be raised as an error any more.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Not tested

Closes #36413 from EnricoMi/branch-ci-change-fail-on-missing-test-files.

Authored-by: Enrico Minack <github@enrico.minack.dev>
Signed-off-by: Sean Owen <srowen@gmail.com>