[SPARK-39714][PYTHON] Try to fix the mypy annotation tests
This patch is for resolving the mypy annotation tests. For making the
 annotation tests, we figure out all failure tests and fix them.

### What changes were proposed in this pull request?

The changed files are all the errors raised when we execute dev/lint-python. All is related mypy annotation test failure.

### Why are the changes needed?

We need to resolve them for avoiding break the common commit merge. This might improve the developer experience for the first insight of pyspark.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Once the PR merged, users/developers can exec `dev/lint-python` for testing.

Closes #37117 from bzhaoopenstack/mypy-annotations-test.

Lead-authored-by: bzhaoop <270162781@qq.com>
Co-authored-by: bzhao <bzhaojyathousandy@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>