[SPARK-39319][CORE][SQL] Make query contexts as a part of `SparkThrowable`
### What changes were proposed in this pull request?
In the PR, I propose to add new interface `QueryContext` Spark core, and allow to get an instance of `QueryContext` from Spark's exceptions of the type `SparkThrowable`. For instance, `QueryContext` should help users to figure out where an error occur while executing queries in Spark SQL.

Also this PR adds `SqlQueryContext` as one of implementation of `QueryContext` to Spark SQL `Origin` which contains a context of TreeNodes + textual summary of the error. The `context` value in `Origin` will have all necessary structural info about the fragment of SQL query to which an error can be linked.

All Spark's exceptions are modified to accept the optional `QueryContext` and pre-built text summary. Apparently, SQL expressions init and pass new context to exceptions.

Closes #36702

### Why are the changes needed?
In the future, this enriches the information of the error message. With the change, it is possible to have a new pretty printing format error message like
```sql
> SELECT * FROM v1;

{
  “errorClass” : [ “DIVIDE_BY_ZERO” ],
  “parameters” : [ {
                   “name” = “config”,
                   “value” = “spark.sql.ansi.enabled”
                 }
                 ],
  “sqlState” : “42000”,
  “context” : {
      “objectType” : “VIEW”,
      “objectName” : “default.v1”
      “indexStart” : 36,
      “indexEnd” : 41,
      “fragment” : “1 / 0” }
   }
}
```

### Does this PR introduce _any_ user-facing change?
Yes. The PR changes Spark's exception by replacing the type of `queryContext` from `String` to `Option[QueryContext]`. User's code can fail if it uses `queryContext`.

### How was this patch tested?
By running the modified test suites:
```
$ build/sbt "test:testOnly *DecimalExpressionSuite"
$ build/sbt "test:testOnly *TreeNodeSuite"
```
and affected test suites:
```
$ build/sbt "sql/testOnly org.apache.spark.sql.SQLQueryTestSuite"
```

Authored-by: Max Gekk <max.gekkgmail.com>
Co-authored-by: Gengliang Wang <gengliangapache.org>

Closes #37209 from MaxGekk/query-context-in-sparkthrowable.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>