[SPARK-40245][SQL] Fix FileScan equality check when partition or data filter columns are not read
### What changes were proposed in this pull request?
Unfortunately the fix in https://github.com/apache/spark/pull/31848 was not correct in all cases. When the partition or data filter contains a column that is not in `readSchema()` the filter nornalization in `FileScan.equals()` doesn't work.

### Why are the changes needed?
To fix `FileScan.equals()` to fix reuse issues.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Added new UT.

Closes #37693 from peter-toth/SPARK-40245-fix-filescan-equals.

Authored-by: Peter Toth <ptoth@cloudera.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>