[SPARK-40295][SQL] Allow v2 functions with literal args in write distribution/ordering
### What changes were proposed in this pull request?

This PR adapts `V2ExpressionUtils` to support arbitrary transforms with multiple args that are either references or literals.

### Why are the changes needed?

After PR #36995, data sources can request distribution and ordering that reference v2 functions. If a data source needs a transform with multiple input args or a transform where not all args are references, Spark will throw an exception.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This PR adapts the test added recently in PR #36995.

Closes #37749 from aokolnychyi/spark-40295.

Lead-authored-by: aokolnychyi <aokolnychyi@apple.com>
Co-authored-by: Anton Okolnychyi <aokolnychyi@apple.com>
Signed-off-by: Chao Sun <sunchao@apple.com>