[SPARK-40280][SQL] Add support for parquet push down for annotated int and long
### What changes were proposed in this pull request?
This fixes SPARK-40280 by normalizing a parquet int/long that has optional metadata with it to look like the expected version that does not have the extra metadata.

## Why are the changes needed?
This allows predicate push down in parquet to work when reading files that are complaint with the parquet specification, but different from what Spark writes.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
I added unit tests that cover this use case. I also did some manual testing on some queries to verify that less data is actually read after this change.

Closes #37747 from revans2/normalize_int_long_parquet_push.

Authored-by: Robert (Bobby) Evans <bobby@apache.org>
Signed-off-by: Thomas Graves <tgraves@apache.org>