[SPARK-40492][SS] Do maintenance before streaming StateStore unload
### What changes were proposed in this pull request?
Before unload of a StateStore, perform a cleanup.

### Why are the changes needed?
Current the maintenance of StateStore is performed by a periodic task in the management thread. If a streaming query become inactive before the next maintenance task fire, its StateStore will be unloaded before cleanup.
There are 2 cases when a StateStore is unloaded.
1. StateStoreProvider is not longer active in the system, for example, when a query ends or the spark context terminates.
2. There is other active StateStoreProvider in the system, for example, when a partition is reassigned.

In case 1, we should do one last maintenance before unloading the instance,

### Does this PR introduce _any_ user-facing change?
Shutdown delay of a query may increase because a maintenance task is scheduled.

### How was this patch tested?
Add an integration test that verify that redundant delta file is deleted when StateStore instances is deactivated and unloaded.

Closes #37935 from chaoqin-li1123/statestore_cleanup.

Lead-authored-by: Chaoqin Li <chaoqin.li@databricks.com>
Co-authored-by: chaoqin-li1123 <55518381+chaoqin-li1123@users.noreply.github.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>