[SPARK-41042][SQL] Rename `PARSE_CHAR_MISSING_LENGTH` to `DATATYPE_MISSING_SIZE`
### What changes were proposed in this pull request?

This PR proposes to rename `PARSE_CHAR_MISSING_LENGTH` to `DATATYPE_MISSING_SIZE`

### Why are the changes needed?

The error class name is specific to `CHAR`, but we get <type> as a parameter in this error class.

The error class should cover more general and have clear name, such as the future types beyond `CHAR`.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

```
./build/sbt “sql/testOnly org.apache.spark.sql.SQLQueryTestSuite*”
```

Closes #38554 from itholic/SPARK-41042.

Authored-by: itholic <haejoon.lee@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>