[SPARK-42790][SQL] Abstract the excluded method for better test for JDBC docker tests
### What changes were proposed in this pull request?
Currently, The JDBC V2 test cases create a lot of method that tests some syntax or function. It is very easy missing the test cases if developer forget to call it.

This PR let all the JDBC V2 could execute these tests by default. If one database doesn't support the related syntax of function, developer have to add the name of this test case to excluded list. We can ensure that use cases are tested.

### Why are the changes needed?
This PR abstract the excluded method for better test for JDBC docker tests.

### Does this PR introduce _any_ user-facing change?
'No'.
Just related to tests.

### How was this patch tested?
Just related to tests.

Closes #40418 from beliefer/SPARK-42790.

Authored-by: Jiaan Geng <beliefer@163.com>
Signed-off-by: Sean Owen <srowen@gmail.com>