Revert [SPARK-50215][SQL] Refactored StringType pattern matching in jdbc code stack
### What changes were proposed in this pull request?

I propose reverting the PR for changing pattern matching of `StringType` in the jdbc code stack, since it may lead to collated column being mapped to uncollated column in some dialects. For the time being, this is not the correct behavior.

### Why are the changes needed?

These changes are needed in order to preserve proper behavior in the dialects regarding datatype mapping.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

No testing was needed.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48833 from vladanvasi-db/vladanvasi-db/jdbc-refactor-revert.

Authored-by: Vladan Vasić <vladan.vasic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>