[MINOR] Fix some typos
### What changes were proposed in this pull request?
This PR fixes typos in docstrings and comments (and a few functional bits of code). The typos were found using the typos software: https://github.com/crate-ci/typos

I've made a typo-fix-PR before, but haven't seen the result on the website yet. Is there anything else I need to do for that?

### Why are the changes needed?
Nice to fix :)

### Does this PR introduce _any_ user-facing change?
yes, documentation was updated.

### How was this patch tested?
No tests added.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #48557 from KazMiddelhoek/master.

Lead-authored-by: Kaz <kaz.middelhoek@ns.nl>
Co-authored-by: Kaz <62500382+KazMiddelhoek@users.noreply.github.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>