[SPARK-50666][SQL] Support hint for reading in JDBC data source
### What changes were proposed in this pull request?

This PR aims to add a hint option for JDBC data source. This option is used to specify the hint for reading. It will apply only if the underlying DBMS supports the hint feature. Currently, this option is only supported by OracleDialect and MySQLDialect.

### Why are the changes needed?

It's useful for performance tuning when reading from DBMS.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Passed GA and add a new test case.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49564 from wayneguow/jdbc_hint.

Authored-by: Wei Guo <guow93@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>