[SPARK-42746][FOLLOW-UP][SQL] Fix flaky listagg tests and refactor code
### What changes were proposed in this pull request?
It was noticed that some tests could be flaky for listagg. The flakiness mainly can come due to non-deterministic behaviour of grouping by a column or collations (i.e. utf8_lcase, unicode_rtrim...). This PR aims to make queries more deterministic by checking only the required info in the result of listagg queries.

### Why are the changes needed?
We do not want to waste resources on flaky tests. Also, we want the tests to scope the minimum set of truth,  which means we do not assume ordering in specific result, when there was an aggregation or grouping.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Test only change.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #50338 from mihailom-db/listaggFollowUp.

Authored-by: Mihailo Milosevic <mihailo.milosevic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>