[SPARK-53785][SS] Memory Source for RTM
### What changes were proposed in this pull request?

Add a memory source implementation that support Real-time Mode.  This source is going to be used to test Real-time Mode.  In this implementation of the memory source, a RPC server is set up on the driver and source tasks will constantly pull this RPC server for new data.  This differs from the existing memory source implementation as data is sent once to tasks as part of the Partition/Split metadata at the beginning of a batch.

### Why are the changes needed?

To test Real-time Mode queries.

### Does this PR introduce _any_ user-facing change?

No, this source is purely to help testing.

### How was this patch tested?

Actual unit tests to test RTM will be added in the future once the engine supports actually running queries in RTM.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52502 from jerrypeng/SPARK-53785.

Authored-by: Jerry Peng <jerry.peng@databricks.com>
Signed-off-by: Liang-Chi Hsieh <viirya@gmail.com>